requirements.txt
Django
pymongo==3.12.3
djongo==1.3.6
pytz
dnspython
djangorestframework
django-cors-headers
psycopg2
python-dateutil
geopy
requests
djangorestframework-simplejwt
cloudinary